feat: 하단 네비게이션 바 스크린 추가
*   feat: 물건 등록 스크린 추가
*   feat: 채팅 스크린 추가
*   feat: 거래 스크린 추가
*   feat: 유저 정보 스크린 추가
* refactor: 하단 네비게이션 바 각 화면들을 클래스화 하여 분리
