+++ app/src/main/java/com/example/swap_it/ui/add_item/AddItemScreen.kt
@@ -1,0 +1,34 @@
+package com.example.swap_it.ui.add_item
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+
+class AddItemScreen {
+
+    @Composable
+    fun AddItemScreen() {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(Color.Yellow)
+        ) {
+            Text(
+                text = "물건 등록",
+                style = MaterialTheme.typography.labelLarge,
+                textAlign = TextAlign.Center,
+                color = Color.White,
+                modifier = Modifier.align(Alignment.Center)
+            )
+        }
+    }
+
+
+}
