+++ app/src/main/java/com/example/swap_it/ui/shopping_list/ShoppingListScreen.kt
@@ -1,0 +1,33 @@
+package com.example.swap_it.ui.shopping_list
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Color.Companion.Gray
+import androidx.compose.ui.text.style.TextAlign
+
+class ShoppingListScreen {
+    @Composable
+    fun ShoppingListScreen() {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(Gray)
+        ) {
+            Text(
+                text = "거래",
+                style = MaterialTheme.typography.labelLarge,
+                textAlign = TextAlign.Center,
+                color = Color.White,
+                modifier = Modifier.align(Alignment.Center)
+            )
+        }
+    }
+
+}
